changed folder structure, separated style files
