+++ App/frontend/src/App.css
@@ -1,63 +1,12 @@
-.App {
+/*.App {
   background-color: black;
   background-image: url("https://i.pinimg.com/564x/87/0e/a9/870ea9ac4fef31c40b2253c6a6fe1b10.jpg");
   background-size: cover;
-}
-
-/*.form-container {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  width: 100vw;
-  height: 70vh;
 }*/
-
-/* Form Related */
-
-.form {
-  position: fixed;
-  z-index: 2;
-  top: 20%;
-  left: calc(50% - 210px); /*%50 - width/2 */
-  width: 420px;
-  box-shadow: rgb(0 0 0 / 16%) 1px 1px 10px;
-  padding-top: 30px;
-  padding-bottom: 20px;
-  border-radius: 8px;
-  background-color: white;
-}
-
-.form-content {
-  padding-left: 12%;
-  padding-right: 12%;
-}
-
-.form-title {
-  text-align: center;
-  font-family: "Inter", sans-serif;
-  margin-bottom: 1em;
-  font-size: 24px;
-  color: rgb(34, 34, 34);
-  font-weight: 800;
-}
 
 .form-control:focus {
   border-color: #e63946 !important;
   box-shadow: none !important;
-}
-
-.label {
-  font-size: 14px;
-  font-weight: 600;
-  color: rgb(34, 34, 34);
-}
-
-.wrapper {
-  /*background-color: #eeebdd;*/
-  font-family: "Inter", sans-serif;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
 }
 
 /* Primary Button */
@@ -87,7 +36,6 @@
   background-color: transparent !important;
   border-color: transparent !important;
 }
-
 
 /* Backgrounds */
 .bg-primary {
