+++ App/frontend/src/components/Backdrop.js
@@ -1,4 +1,4 @@
-import "./Backdrop.css";
+import "./styles/Backdrop.css";
 
 function Backdrop(props) {
   return <div className="backdrop" onClick={props.onClick} />;
