+++ App/frontend/src/components/MenuButton.js
@@ -1,8 +1,8 @@
 import Button from "react-bootstrap/Button";
-import { AiOutlineMenu } from "react-icons/ai";
+// import { AiOutlineMenu } from "react-icons/ai";
 import { SlMenu } from "react-icons/sl";
 import { useState } from "react";
-import "./MenuButton.css";
+import "./styles/MenuButton.css";
 
 function MenuButton() {
   const [rotateMenuButton, setRotateMenuButton] = useState(false);
