+++ App/frontend/src/pages/styles/Access.css
@@ -1,0 +1,40 @@
+/*.access-container {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  width: 100vw;
+  height: 70vh;
+}*/
+
+.access {
+  position: fixed;
+  z-index: 2;
+  top: 20%;
+  left: calc(50% - 210px); /*%50 - width/2 */
+  width: 420px;
+  box-shadow: rgb(0 0 0 / 16%) 1px 1px 10px;
+  padding-top: 30px;
+  padding-bottom: 20px;
+  border-radius: 8px;
+  background-color: white;
+}
+
+.access-content {
+  padding-left: 12%;
+  padding-right: 12%;
+}
+
+.access-title {
+  text-align: center;
+  font-family: "Inter", sans-serif;
+  margin-bottom: 1em;
+  font-size: 24px;
+  color: rgb(34, 34, 34);
+  font-weight: 800;
+}
+
+.access-label {
+  font-size: 14px;
+  font-weight: 600;
+  color: rgb(34, 34, 34);
+}
